Conveyor Belt functionality
Script that causes objects to move across the surface of another object (Conveyor Belt). Drag and drop functionality - Object returns to the location it was picked up at if its dropped in an invalid location. Drag and drop script to be added to Trash objects to allow them to be picked up by the player. Conveyor belt script to be added to the conveyor belt object. Trash items must have "MoveableObject" tag applied to them for full functionality.  - MAT