Fix Drag and Drop + Bin Animations
Fixed the movement of the 3d trash items to follow a consistent path and make it easier to navigate the correct bin

Added bin animations to indicate  which bin is chosen by the user whilst hovering over it.

(Note: Fixed an issue with the tags, the types of bins need to be re-adjusted again, due to removing their tags. This is due to keeping the code consistent with the colliders, but also because they needed to be added into the "layers" instead of the "tags")