Added some randomness.
There is now a 1 in 10 chance that 2 items spawn on the conveyor belt at once and a very low chance that 3 items spawn.  I've made it so the trash interval can not decrease below 4 seconds and also there is now a function to stop the conveyor belt for 5 seconds, currently the hot key for this is "J".